add docker2fl pkg and bin
