+++ Cargo.toml
@@ -7,6 +7,18 @@ edition = "2021"
 [build-dependencies]
 git-version = "0.3.5"
 
+[lib]
+name = "docker2fl"
+path = "src/docker2fl.rs"
+
+[[bin]]
+name = "docker2fl"
+path = "src/main.rs"
+
+[[bin]]
+name = "fl-server"
+path = "src/server/main.rs"
+
 [dependencies]
 log = "0.4"
 anyhow = "1.0.44"
@@ -21,6 +33,11 @@ tempdir = "0.3"
 serde_json = "1.0"
 toml = "0.4.2"
 clap = { version = "4.2", features = ["derive"] }
+
+serde = { version = "1.0.159" , features = ["derive"] }
+axum = "0.7.3"
+chrono = { version = "0.4", features = ["serde"] }
+tower-http = { version = "0.5.0", features = ["fs", "cors"] }
 
 [profile.release]
 lto = true
